backend/requirements.txt
Flask
pycryptodome
hvac
